+++ src/index.js
@@ -1,4 +1,0 @@
-import FileUpload from "./FileUpload.jsx"
-
-export default FileUpload
-