+++ src/index.ts
@@ -1,0 +1,4 @@
+import FileUpload from "./FileUpload.js"
+
+export default FileUpload
+